Project Files/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
streamlit
httpx
pandas
scikit-learn
pinecone-client
sentence-transformers
markdown
ibm-watsonx
